requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2
django-background-tasks==1.2.5
django-compat==1.0.15
django-crispy-forms==1.12.0
django-filter==2.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
idna==3.2
Markdown==3.3.4
mysqlclient==2.0.3
Pillow==8.3.1
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.7
